// src/lib/generated/requisition/models/paymentTerms.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Kaizen Admin Approval System
 * Kaizen Admin and Approval Management System using LangGraph
 * OpenAPI spec version: 1.0.0
 */

export type PaymentTerms = typeof PaymentTerms[keyof typeof PaymentTerms];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PaymentTerms = {
  net_15: 'net_15',
  net_30: 'net_30',
  net_45: 'net_45',
  net_60: 'net_60',
  net_90: 'net_90',
  cod: 'cod',
  prepaid: 'prepaid',
  due_on_receipt: 'due_on_receipt',
  custom: 'custom',
} as const;